Move validators to helpers folder
