+++ packages/kbn-form-lib/src/helpers/field_validators/empty_field.ts
@@ -7,8 +7,8 @@
  */
 
 import { ValidationFunc } from '../../hook_form_lib';
-import { isEmptyString } from '../../validators/string';
-import { isEmptyArray } from '../../validators/array';
+import { isEmptyString } from '../validators/string';
+import { isEmptyArray } from '../validators/array';
 import { ERROR_CODE } from './types';
 
 export const emptyField =
