+++ packages/kbn-form-lib/src/helpers/field_validators/index_pattern_field.ts
@@ -7,7 +7,7 @@
  */
 
 import { ValidationFunc } from '../../hook_form_lib';
-import { containsChars } from '../../validators/string';
+import { containsChars } from '../validators/string';
 import { ERROR_CODE } from './types';
 
 // --- start copy from Dataview plugin
