+++ packages/kbn-form-lib/src/helpers/field_validators/lowercase_string.ts
@@ -7,7 +7,7 @@
  */
 
 import { ValidationFunc } from '../../hook_form_lib';
-import { isLowerCaseString } from '../../validators/string';
+import { isLowerCaseString } from '../validators/string';
 import { ERROR_CODE } from './types';
 
 export const lowerCaseStringField =
