+++ packages/kbn-form-lib/src/helpers/field_validators/min_selectable_selection.ts
@@ -9,7 +9,7 @@
 import { EuiSelectableOption } from '@elastic/eui';
 
 import { ValidationFunc, ValidationError } from '../../hook_form_lib';
-import { hasMinLengthArray } from '../../validators/array';
+import { hasMinLengthArray } from '../validators/array';
 import { multiSelectComponent } from '../serializers';
 import { ERROR_CODE } from './types';
