+++ packages/kbn-form-lib/src/helpers/field_validators/number_smaller_than.ts
@@ -7,7 +7,7 @@
  */
 
 import { ValidationFunc, ValidationError } from '../../hook_form_lib';
-import { isNumberSmallerThan } from '../../validators/number';
+import { isNumberSmallerThan } from '../validators/number';
 import { ERROR_CODE } from './types';
 
 export const numberSmallerThanField =
