+++ packages/kbn-form-lib/src/helpers/field_validators/starts_with.ts
@@ -7,7 +7,7 @@
  */
 
 import { ValidationFunc } from '../../hook_form_lib';
-import { startsWith } from '../../validators/string';
+import { startsWith } from '../validators/string';
 import { ERROR_CODE } from './types';
 
 export const startsWithField =
