+++ packages/kbn-form-lib/src/helpers/validators/array/is_empty.ts
