+++ packages/kbn-form-lib/src/helpers/validators/string/contains_chars.ts
