+++ packages/kbn-form-lib/src/helpers/validators/string/has_min_length.ts
