+++ packages/kbn-form-lib/src/helpers/validators/string/is_empty.ts
